refactor(064): Extract PII sanitization to shared/lib/privacy.ts for reuse
Bot review feedback: Extract PII patterns for cross-feature reuse

Changes:
- Create shared/lib/privacy.ts with PII sanitization logic
- Move PII_PATTERNS, PII_STOP_WORDS, sanitizePII() from features/goals/lib/export.ts
- Update export.ts to import from @/shared/lib/privacy
- Comprehensive JSDoc for reusability across features

Benefits:
- Single source of truth for privacy patterns (DRY principle)
- Reusable across transactions, budgets, categories features
- Constitution Principle I (Privacy-First) compliance

Deferred to Linear:
- MMT-116: Memoize storage quota check (Phase 2 performance)
- MMT-117: International phone support (Phase 2+, when users request)
