+++ frontend/src/features/goals/lib/export.ts
@@ -10,143 +10,7 @@
 import Papa from 'papaparse';
 import type { Goal } from '../types/goal';
 import type { Contribution } from '../types/contribution';
-
-/**
- * PII Sanitization Patterns (Constitution Requirement)
- *
- * Regex patterns with word boundaries to detect and sanitize:
- * - Email addresses
- * - Phone numbers (US format)
- * - Social Security Numbers (SSN)
- * - Credit card numbers (13-19 digits with optional spaces/dashes)
- * - Names (common patterns like "John Doe", "Jane Smith")
- * - Street addresses (numbers + street names)
- */
-const PII_PATTERNS = {
-  // Email: user@domain.com
-  email: /\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b/g,
-
-  // Phone: (123) 456-7890, 123-456-7890, +1-555-123-4567
-  // Bot review M2: International Phone Format Coverage
-  //
-  // **Current Scope**: US phone numbers only (+1 country code, 10-digit format)
-  // **Matches**: (555) 123-4567, 555-123-4567, +1-555-123-4567, 1-555-123-4567
-  // **Does NOT match**: International formats (e.g., +44, +61, +33, +49, +86, +91)
-  //
-  // **Rationale for US-only in Phase 1**:
-  // - Conservative approach: Avoids false positives on non-phone numbers
-  // - Goal names may contain digit patterns (e.g., "Save $1,234 for vacation")
-  // - International regex patterns are complex and error-prone (many country-specific rules)
-  //
-  // **Known Limitations** (documented for Phase 2):
-  // 1. UK (+44): +44 20 7946 0958, +44 7911 123456 (landline vs mobile)
-  // 2. Australia (+61): +61 2 1234 5678, +61 412 345 678 (landline vs mobile)
-  // 3. France (+33): +33 1 23 45 67 89 (space-separated groups)
-  // 4. Germany (+49): +49 30 123456, +49 151 12345678 (variable length)
-  // 5. China (+86): +86 10 1234 5678, +86 138 1234 5678 (variable length)
-  // 6. India (+91): +91 11 2345 6789, +91 98765 43210 (variable length)
-  //
-  // **Future Enhancement** (Phase 2):
-  // - Use libphonenumber-js library for comprehensive international support
-  // - Support 200+ country codes with format validation
-  // - Handle variable-length numbers (7-15 digits per ITU-T E.164 standard)
-  //
-  // **Alternative considered**: Broad regex matching +[0-9]{1,3}[-.\s]?[0-9]{7,15}
-  // **Rejected**: Too many false positives (dates, amounts, goal names with numbers)
-  phone: /(?:\+?1[-.\s]?)?\(?[0-9]{3}\)?[-.\s]?[0-9]{3}[-.\s]?[0-9]{4}\b/g,
-
-  // SSN: 123-45-6789, 123456789
-  ssn: /\b\d{3}[-\s]?\d{2}[-\s]?\d{4}\b/g,
-
-  // Credit Card: 1234-5678-9012-3456 (13-19 digits, Amex 15 digits supported)
-  // Bot review H1: Added patterns for no-separator formats
-  // Patterns: 16-digit no spaces (1234567890123456), 15-digit Amex (123456789012345),
-  //           4-6-5 Amex with separators, 4-4-4-4 standard with separators
-  creditCard: /\b(?:\d{16}|\d{15}|\d{4}[-\s]?\d{6}[-\s]?\d{5}|\d{4}[-\s]?\d{4}[-\s]?\d{4}[-\s]?\d{3,7})\b/g,
-
-  // Names: "John Doe", "Jane Smith" (capitalized first/last name pattern)
-  // PR86-3: Restored with word-boundary matching (Feature 019 approach)
-  // Matches: capitalized First Last (e.g., "Sarah Johnson", "Michael Chen")
-  // Stop-word filtering prevents false positives on goal names (see GOAL_NAME_STOP_WORDS)
-  name: /\b[A-Z][a-z]+\s+[A-Z][a-z]+\b/g,
-
-  // Street Address: "123 Main St", "456 Oak Avenue"
-  // Bot review C2: Require street name IMMEDIATELY before street type
-  // Pattern: <number> <name> <type> (no extra words allowed between name and type)
-  address: /\b\d{1,5}\s+[A-Za-z0-9]+\s+(?:Street|St|Avenue|Ave|Road|Rd|Drive|Dr|Lane|Ln|Boulevard|Blvd)\b/gi,
-};
-
-/**
- * Stop-words for name sanitization (PR86-3)
- *
- * Common goal name terms that should NOT be redacted as names.
- * Prevents false positives on goal names like "Emergency Fund" or "House Payment"
- *
- * Examples:
- * - "Emergency Fund" → NOT redacted (contains "Emergency" and "Fund")
- * - "John Smith" → REDACTED (no stop-words)
- * - "Car Payment" → NOT redacted (contains "Car" and "Payment")
- */
-const GOAL_NAME_STOP_WORDS = new Set([
-  'Emergency',
-  'House',
-  'Car',
-  'Wedding',
-  'Vacation',
-  'Retirement',
-  'College',
-  'Fund',
-  'Savings',
-  'Account',
-  'Budget',
-  'Payment',
-  'Down',
-  'Trip',
-  'Goal',
-  // Common street/location words that might appear in goal names
-  // PR86-3: Prevents false positives on location names like "Main Street" or "Oak Avenue"
-  'Main',
-  'Street',
-  'Oak',
-  'Pine',
-  'Elm',
-  'Maple',
-  'Avenue',
-  'Road',
-  'Drive',
-  'Mall', // Shopping centers
-  'Visit', // Travel-related
-]);
-
-/**
- * Sanitize text by replacing PII with redacted placeholders
- *
- * @param text - Text to sanitize
- * @returns Sanitized text with PII replaced
- */
-export function sanitizePII(text: string): string {
-  let sanitized = text;
-
-  // Replace each PII pattern with [REDACTED] placeholder
-  // Note: Order matters! Process addresses before names to avoid false positives
-  sanitized = sanitized.replace(PII_PATTERNS.email, '[EMAIL_REDACTED]');
-  sanitized = sanitized.replace(PII_PATTERNS.creditCard, '[CARD_REDACTED]');
-  sanitized = sanitized.replace(PII_PATTERNS.ssn, '[SSN_REDACTED]');
-  sanitized = sanitized.replace(PII_PATTERNS.phone, '[PHONE_REDACTED]');
-  sanitized = sanitized.replace(PII_PATTERNS.address, '[ADDRESS_REDACTED]');
-
-  // PR86-3: Name sanitization with stop-word filtering (restored from PR85-6 removal)
-  // Check each name match against stop-words to avoid false positives on goal names
-  sanitized = sanitized.replace(PII_PATTERNS.name, (match) => {
-    // Check if match contains any stop-word (case-insensitive)
-    const hasStopWord = Array.from(GOAL_NAME_STOP_WORDS).some((word) =>
-      match.includes(word)
-    );
-    return hasStopWord ? match : '[NAME_REDACTED]';
-  });
-
-  return sanitized;
-}
+import { sanitizePII } from '@/shared/lib/privacy';
 
 /**
  * Format goal for export with PII sanitization
